requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Django==2.2.2
django-filter==2.1.0
djangorestframework==3.9.4
httpie==1.0.2
idna==2.8
Jinja2==2.10.1
lxml==4.3.4
Markdown==3.1.1
MarkupSafe==1.1.1
nltk==3.4.3
numpy==1.16.4
pandas==0.24.2
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
